fixed some problems

+++ src/RaqamliAvlod.Infrastructure.Service/Services/Courses/CourseCommentService.cs
@@ -8,34 +8,30 @@ using RaqamliAvlod.Infrastructure.Service.Dtos;
 using RaqamliAvlod.Infrastructure.Service.Helpers;
 using RaqamliAvlod.Infrastructure.Service.Interfaces.Common;
 using RaqamliAvlod.Infrastructure.Service.Interfaces.Courses;
-using RaqamliAvlod.Infrastructure.Service.Services.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RaqamliAvlod.Infrastructure.Service.Services.Courses;
 
 public class CourseCommentService : ICourseCommentService
 {
     private readonly IUnitOfWork _unitOfWork;
+    private readonly IPaginatorService _paginator;
 
-    public CourseCommentService(IUnitOfWork unitOfWork)
+    public CourseCommentService(IUnitOfWork unitOfWork, IPaginatorService paginatorService, IIdentityHelperService identityHelper)
     {
         _unitOfWork = unitOfWork;
+        _paginator = paginatorService;
     }
-    public async Task<bool> CreateAsync(long courseId, CourseCommentCreateDto dto)
+    public async Task<bool> CreateAsync(long userId, long courseId, CourseCommentCreateDto dto)
     {
-        var user = await _unitOfWork.Users.FindByIdAsync(dto.OwnerId);
+        var user = await _unitOfWork.Users.FindByIdAsync(userId);
         var course = await _unitOfWork.Courses.FindByIdAsync(courseId);
 
         if (user is null)
             throw new StatusCodeException(HttpStatusCode.BadRequest, "Owner not found!");
         if (course is null)
             throw new StatusCodeException(HttpStatusCode.BadRequest, "Course not found!");
-        
+
         var courseComment = (CourseComment)dto;
         courseComment.CourseId = courseId;
         courseComment.CreatedAt = TimeHelper.GetCurrentDateTime();
@@ -45,17 +41,19 @@ public class CourseCommentService : ICourseCommentService
 
     }
 
-    public async Task<bool> DeleteAsync(long courseId, long id)
+    public async Task<bool> DeleteAsync(long userId, long courseId, long id)
     {
         var course = await _unitOfWork.Courses.FindByIdAsync(courseId);
         var courseComment = await _unitOfWork.CourseComments.FindByIdAsync(id);
-        
+
         if (course is null)
             throw new StatusCodeException(HttpStatusCode.BadRequest, "Course not found!");
         if (courseComment is null)
-            throw new StatusCodeException(HttpStatusCode.BadRequest, "Course comment not found!");
-        if (courseComment.CourseId!=courseId)
-            throw new StatusCodeException(HttpStatusCode.BadRequest, "Comment not found in this course!");
+            throw new StatusCodeException(HttpStatusCode.BadRequest, "Comment not found!");
+        if (courseComment.CourseId != courseId)
+            throw new StatusCodeException(HttpStatusCode.BadRequest, "This comment related to another course");
+        if (courseComment.OwnerId != userId)
+            throw new StatusCodeException(HttpStatusCode.BadRequest, "Not permitted!");
         var res = await _unitOfWork.CourseComments.DeleteAsync(id);
 
         return res is not null ? true : false;
@@ -64,16 +62,17 @@ public class CourseCommentService : ICourseCommentService
     public async Task<IEnumerable<CourseCommentViewModel>> GetAllByCourseIdAsync(long courseId, PaginationParams @params)
     {
         var courseComments = await _unitOfWork.CourseComments.GetAllByCourseIdAsync(courseId, @params);
+        _paginator.ToPagenator(courseComments.MetaData);
 
         var courseCommentViews = new List<CourseCommentViewModel>();
 
         foreach (var courseComment in courseComments)
         {
-            var owner  = (await _unitOfWork.Users.FindByIdAsync(courseComment.OwnerId))!;
+            var owner = (await _unitOfWork.Users.FindByIdAsync(courseComment.OwnerId))!;
 
             var courseCommentView = (CourseCommentViewModel)courseComment;
 
-            courseCommentView.Username = owner.Username!;
+            courseCommentView.Owner = (OwnerViewModel)owner;
 
             courseCommentViews.Add(courseCommentView);
         }
@@ -95,7 +94,7 @@ public class CourseCommentService : ICourseCommentService
 
         var courseCommentView = (CourseCommentViewModel)courseComment;
 
-        courseCommentView.Username = owner.Username!;
+        courseCommentView.Owner = (OwnerViewModel)owner;
 
         return courseCommentView;
     }
